+++ src/main/scala/org/spartanz/parserz/Enumerable.scala
@@ -1,0 +1,8 @@
+package org.spartanz.parserz
+
+trait Enumerable[A] {
+
+  val min: A
+  val max: A
+  def range(min: A, max: A): Iterable[A]
+}
